updates to deploy to heroku

+++ db/migrate/20140527084124_add_x_to_entities.rb
@@ -1,5 +1,5 @@
 class AddXToEntities < ActiveRecord::Migration
   def change
-    add_column :entities, :x, :double
+    add_column :entities, :x, :integer
   end
 end
